refactor(spaces): update import paths for poll types in stories
- Adjusted import paths to use `poll-question` type from the new module
  `@/features/spaces/polls/types`
- Updated stories for `report`, `dropdown-viewer`, and `linear-scale-viewer`
