+++ ts-packages/web/src/features/spaces/components/survey/viewer/linear-scale-viewer.stories.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import LinearScaleViewer from './linear-scale-viewer';
-import { SurveyAnswerType } from '../../../../../types/survey-type';
+import { SurveyAnswerType } from '@/features/spaces/polls/types/poll-question';
+
 const mockT = ((key: string) => key) as any;
 
 export default {
